separated RedisCluster from client
